chore(server): rename rest module to api
