+++ packages/botonic-server/src/index.ts
@@ -1,8 +1,8 @@
+export * from './api'
 export * from './constants'
 export * from './data-provider'
 export * from './dispatchers'
 export * from './handlers'
 export * from './notifying'
-export * from './rest'
 export * from './sqs-utils'
 export * from './websocket'
