+++ packages/create-botonic-app/dev-template/api/src/rest/app.ts
@@ -1,4 +1,4 @@
-import { routes } from '@botonic/server/src/rest/routes'
+import { routes } from '@botonic/server/src/api/routes'
 import * as cors from 'cors'
 import * as express from 'express'
 import * as morgan from 'morgan'
